+++ lib/tasks/docs.rake
@@ -68,7 +68,7 @@ namespace :docs do
       Primer::LayoutComponent,
       Primer::Beta::Link,
       Primer::Markdown,
-      Primer::MenuComponent,
+      Primer::Alpha::Menu,
       Primer::Navigation::TabComponent,
       Primer::OcticonComponent,
       Primer::Beta::Popover,
