+++ previews/primer/alpha/menu_preview.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Primer
+  module Alpha
+    # @label Menu
+    class MenuPreview < ViewComponent::Preview
+      # @label Playground
+      def playground; end
+
+      # @label Default
+      def default; end
+    end
+  end
+end
