--- previews/primer/menu_component_preview.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-module Primer
-  # @label MenuComponent
-  class MenuComponentPreview < ViewComponent::Preview
-    # @label Playground
-    def playground; end
-
-    # @label Default
-    def default; end
-  end
-end
