+++ previews/primer/menu_component_preview/playground.html.erb
@@ -1,4 +1,4 @@
-<%= render(Primer::MenuComponent.new) do |c| %>
+<%= render(Primer::Alpha::Menu.new) do |c| %>
   <% c.with_heading(tag: :h2) do %>
     Heading
   <% end %>
