+++ static/audited_at.json
@@ -20,6 +20,7 @@
   "Primer::Alpha::Layout": "",
   "Primer::Alpha::Layout::Main": "",
   "Primer::Alpha::Layout::Sidebar": "",
+  "Primer::Alpha::Menu": "",
   "Primer::Alpha::NavList": "",
   "Primer::Alpha::NavList::Item": "",
   "Primer::Alpha::NavList::Section": "",
