+++ static/statuses.json
@@ -20,6 +20,7 @@
   "Primer::Alpha::Layout": "alpha",
   "Primer::Alpha::Layout::Main": "alpha",
   "Primer::Alpha::Layout::Sidebar": "alpha",
+  "Primer::Alpha::Menu": "alpha",
   "Primer::Alpha::NavList": "alpha",
   "Primer::Alpha::NavList::Item": "alpha",
   "Primer::Alpha::NavList::Section": "alpha",
@@ -78,7 +79,7 @@
   "Primer::LinkComponent": "deprecated",
   "Primer::LocalTime": "alpha",
   "Primer::Markdown": "beta",
-  "Primer::MenuComponent": "alpha",
+  "Primer::MenuComponent": "deprecated",
   "Primer::Navigation::TabComponent": "alpha",
   "Primer::OcticonComponent": "beta",
   "Primer::OcticonSymbolsComponent": "alpha",
