+++ test/components/alpha/menu_test.rb
@@ -2,17 +2,17 @@
 
 require "components/test_helper"
 
-class PrimerMenuComponentTest < Minitest::Test
+class PrimerAlphaMenuTest < Minitest::Test
   include Primer::ComponentTestHelpers
 
   def test_does_not_render_without_items
-    render_inline(Primer::MenuComponent.new)
+    render_inline(Primer::Alpha::Menu.new)
 
     refute_selector(".menu")
   end
 
   def test_renders_items
-    render_inline(Primer::MenuComponent.new) do |c|
+    render_inline(Primer::Alpha::Menu.new) do |c|
       c.item(selected: true, href: "#url") { "Item 1" }
       c.item(href: "#url") { "Item 2" }
     end
@@ -24,7 +24,7 @@ class PrimerMenuComponentTest < Minitest::Test
   end
 
   def test_renders_heading
-    render_inline(Primer::MenuComponent.new) do |c|
+    render_inline(Primer::Alpha::Menu.new) do |c|
       c.heading(tag: :h3) { "Heading" }
       c.item(selected: true, href: "#url") { "Item 1" }
       c.item(href: "#url") { "Item 2" }
@@ -37,7 +37,7 @@ class PrimerMenuComponentTest < Minitest::Test
 
   def test_falls_back_to_h2_when_heading_tag_isnt_valid
     without_fetch_or_fallback_raises do
-      render_inline(Primer::MenuComponent.new) do |c|
+      render_inline(Primer::Alpha::Menu.new) do |c|
         c.heading(tag: :div) { "Heading" }
         c.item(selected: true, href: "#url") { "Item 1" }
         c.item(href: "#url") { "Item 2" }
